// src/app/moduls/offers/model/offer-update.ts
export interface OfferUpdate{
    offerDetailsDto: any;
    id: number,
    kindOfProperty: string;
    price: number;
    title: string;
    city: string;
    street: string;
    numberOfRooms: number;
    area: number;
    pricePerMeter: number;
    floor: number;
    description: string;
    rent: string;
    ownershipForm: string;
    parkingPlace: string;
    finishLevel: string;
    heating: string;
    contactDetails: string;
    market: string;
    announcerType: string;
    yearOfConstruction: number;
    buildingType: string;
    media: string;
    equipment: string;
}